catalog-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# OpenTelemetry instrumentation
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp==1.21.0

# Database (will add PostgreSQL later)
# asyncpg==0.29.0
# sqlalchemy==2.0.23
